+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 #### Added
 
-- [#](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/) Added support for check constraints.
+- [#1141](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1141) Added support for check constraints.
 
 ## v7.1.0.rc2