+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## v7.0.6
 
 #### Added